[FIX/#148] MainFragment 뒤로가기 앱 종료 설정
